global: rename resources 'file' to 'draft_file'
